requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
beautifulsoup4==4.12.2
easyocr==1.7.1
lxml==4.9.3
requests==2.31.0
pandas==2.1.3
python-dotenv==1.0.0
tqdm==4.66.1
matplotlib==3.8.2
numpy==1.26.2
pillow==10.1.0